Refactor smart card code organization
